sprint 27 фиксы верстки
