+++ app/src/main/java/com/example/playlistmaker/settings/ui/SettingsFragment.kt
@@ -27,7 +27,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.net.toUri
@@ -52,7 +51,6 @@ class SettingsFragment : Fragment() {
         }
     }
 
-//    @Preview(device = "id:pixel_8")
     @Composable
     private fun Settings() {
         val darkTheme by viewModel.isDarkTheme.observeAsState(false)
