+++ app/src/main/java/com/example/playlistmaker/uicomponents/TrackItemComponent.kt
@@ -12,14 +12,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.Font
@@ -33,7 +30,7 @@ import com.example.playlistmaker.helpers.TimeHelper.longToMillis
 import com.example.playlistmaker.search.data.TrackDto
 
 @Composable
-fun TrackItem(
+fun TrackItemComponent(
     track: TrackDto,
     onClick: () -> Unit
 ) {
@@ -50,6 +47,8 @@ fun TrackItem(
         AsyncImage(
             model = track.artworkUrl100 ?: R.drawable.track_stub,
             modifier = Modifier.size(48.dp),
+            placeholder = painterResource(R.drawable.track_stub),
+            error = painterResource(R.drawable.track_stub),
             contentDescription = null,
         )
 
@@ -68,7 +67,7 @@ fun TrackItem(
                     maxLines = 1,
                     overflow = TextOverflow.Ellipsis,
                     fontFamily = FontFamily(Font(R.font.ys_display_regular)),
-                    color = colorResource(id = R.color.text_color)
+                    color = colorResource(id = R.color.search_result_track_text_color)
                 )
             }
             Column {
