+++ app/src/main/res/values/colors.xml
@@ -9,6 +9,7 @@
     <color name="white">#FFFFFFFF</color>
     <color name="settings_main_color">#FFFFFFFF</color>
     <color name="main_background">#3772E7</color>
+    <color name="tab_indicator">@color/blue</color>
     <color name="grey">#AEAFB4</color>
     <color name="light_grey">#E6E8EB</color>
     <color name="settings_text_color">#000000</color>
